VT: Improve observable indexing and recovery
This patch modifies the VirtualTopology and the DeviceMapper abstract
class to index observables with a key based on both the device class
and device identifiers. This allows the virtual topology to cache
observables for multiple devices classes and the same device identifier
(e.g. a router and BGP router observable for the same router
identifier).

The patch does not modify the devices cache, which still provides
indexing only based on the device identifier.

Additionaly, the patch revamps the device mapper state and error
handling with the following changes:
 - It increases the device mapper states from two (subscribed/
   unsubscribed) to five, which includes a state (Closed) indicating
   that the mapper is unusable, but the underlying device is still
   available.
 - Does not throw an exception during the OnSubscribe.call if the
   mapper observable has completed/emitted an error or when the mapper
   is closed. Instead, it emits the state to the subscriber.
 - The virtual topology includes an onErrorResumeNext to recover
   an observable for a closed mapper to an observable created with
   a fresh new mapper.

Currently, the Close state for device mappers is unused, but it
shall be used in the future (e.g. for the BGP mapper) or to garbage
collect unused observables from the virtual topology.

Change-Id: Ic3beb8ce20f536175d15a4a509cf10e26df13e65
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>