BGP: Reintroducing the BGP mapper for the new model
This patch re-introduces the BgpMapper as BgpRouterMapper, providing
BGP notifications for a given router. The mapper also integrates with
the VirtualTopology, to provide a single source of BGP updates.

Unlike the previous BgpMapper, the notifications from the
BgpRouterMapper must further be filtered using a BgpPortMapper, which
will consolidate the BgpRouter updates with the state of a port.

Ref: MNA-466

Change-Id: I618a246cc56b272ab982750a4a73c467b399d8fb
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>