Cluster API: fix some L4LB tests, disable others
The fixed ones required changing the the following response codes, in
order to remain consistent with the rest of the API:

- update a pool member with a bad pool id, or member id: 400 (bad req)
  -> 404 (not found)

Disabled tests are marked using the
Assume.assumeFalse(!FuncTest.isCompatApiEnabled()) flag, and will be
fixed and enabled later when L4Lb is completed on V2.  An issue is
created for this purpose (MNA-700).

Ref.: MNA-577
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: Ifb9af60d33816ac04e7f9c481e72cd85ee25ce49