VxGW v2: Flooding Proxy management
This patch provides a reimplementation of the flooding proxy management
code, adapted to the new storage stack.  The fp is now published as a
Tunnel Zone key using the state storage.

The patch introduces a FloodingProxyHerald that can be used outside of
the vxgw service to retrieve the flooding proxies for different tunnel
zones.

Change-Id: Ia4fd7df88f9a9223a1cfbe5f03309f1833030ab1
Signed-off-by: Galo Navarro <galo@midokura.com>