Proto: move VtepBinding into Vtep
This patch makes VtepBinding nested into VTEP, losing the fields used
for id and backrefs.

The motivation for this change is to simplify access from the VxGW
service, as the separation doesn't seem to have a real justification at
this point.  The only reason in favour would be to control size on the
VTEP object.  However, the expected number of bindings per VTEP is not
expected to go over the low 100s.  Additionally, even for large Vtep
messages the rate of updates is not expected to be significant as
binding changes are a rare operation.

Vtep API DTOs are adapted, but the VtepBindingResource is not removed
(this may be done in a differnet patch, pending discussion on the final
APIs.)

Change-Id: I214d9f0d537c113579b6e0312ba2b6eade8d327d
Signed-off-by: Galo Navarro <galo@midokura.com>