Update patch set 6
Patch Set 6: Code-Review-2

I'm going to -2 because this will break all tempest jobs as soon as it's merged, given that both ansible and devstack don't support it. 

Feel free to remove me from the reviewers when this is solved, just to make sure people is aware of the blocker.

Patch-set: 6