Update patch set 8
Patch Set 8:

Found the problem: in the compat API (org.midonet.cluster.rest_api.models), the LoadBalancer and Pool resource should have fields corresponding to the new protobuf back-references (annotated with @ZoomField and @JsonIgnore) and the update() method should copy those fields from an old to a new resource.

Otherwise, the protobuf will loose the backreferences when doing a PUT request via the API.

Let me know if you need a hand with this.

Patch-set: 8