Update patch set 2
Patch Set 2:

Hi Jan. This setup_test_server is going away pretty soon when we switch mmm by sandbox (hopefully in a couple of weeks). Not sure if this patch is necessary. I'll update sandbox to install the latest package from 2.0 series instead of pinning to 2.0.10. Does this sound right to you?

Patch-set: 2