Update patch set 2
Patch Set 2:

Thanks Nicolas: This is an early sketch of a state table storage, to abstract the replicated maps in the API. Of course, it's not complete. I agree that if we find too many differences between the MAC and ARP tables, we can provide distinct base classes. For the API at the moment, this is not needed.

Patch-set: 2
Label: Code-Review=0