State storage: barebone implementation for ARP tables
This patch provides a barebone implementation for legacy ARP state
tables based on DataClient. Its purpose is to abstract the data
client implementation using the StateTableStorage trait, and to
facilitate its use in the compatibility API.

Currently only some methods are implemented, with the remaining
methods implemented as needed, as the legacy storage is being
phased out.

Change-Id: I05226dcfc52ba7c87bda1ac5b9098fc834c58eeb
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>