Update patch set 3
Patch Set 3:

In fact it could be easier than I thought. For instance, Ip4ToMacReplicatedMap has lots of methods in the companion object but the class itself is just a replicated map with proper serialization. So we could go ahead with your solution. Even if there's more refactoring to do, the end solution would be cleaner.

Patch-set: 3