L4LB: VIPs are referenced by pools
The patch changes the v2 topology model for L4LB such that VIPs
are now referenced by pools, rather than the load-balancer. To
accommodate the model change, the patch modifies the following.

On the cluster side: It introduces the new L4LB model, by removing
the direct reference and back-references between load-balancer and
VIPs, and replacing them with references between pools and VIPs,
which have already been introduced in a previous patch. The patch
also updates the compatibility API resources, and minor changes to
the C3PO translators.

On the agent side: The patch rewrites the load-balancer, pool, and
pool health-monitor mappers to consider the changed relatioships
between load-balancers, pools, and VIPs.

The patch also ensures that all L4LB API tests are passing in v2.

Change-Id: I78dddf7187571d39191d639f1bc3ea7283ea4044
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>