Cluster: bootstrap workflow, first iteration
Introduce a midonet-cluster-bootstrap executable that will get installed
along with the midonet-cluster package.  After install, this script can
be used to bootstrap the midonet cluster for both a single or multi node
configuration.

Single node:

    $ sudo midonet-cluster-bootstrap
    MidoNet Cluster Bootstrap workflow
    mido-bootstrap > single_node
    Single node bootstrap workflow.  Use `help` for options.
    bootstrap.single-node> mgmt_ip 192.168.2.41
    This Cluster node's management address will be: 192.168.2.41
    bootstrap.single-node> apply
    Writing configuration to: /etc/midonet-cluster/midonet-cluster.conf
    midonet-cluster start/running, process 5537
    bootstrap.single-node>

As shown above, the operator can chose a single_node install.  The
script will show the default configuration and available addresses, and
allow using a "mgmt_addr" command to set a new management address.  When
the operator is satisfied, "apply" will persist the configuration to the
MidoNet Cluster bootstrap configuration file and start the MidoNet
Cluster service.

For a multi-node install the process is very similar, although it must
be carried out in several nodes.  One of these is designated as the lead
for the bootstrap process:

    bootstrap> multi_node
    Multi node bootstrap workflow.  Use `help`for options.
    bootstrap.multi-node> mgmt_ip 192.168.2.41
    This Cluster node's management address will be: 192.168.2.41
    bootstrap.multi-node> lead_bootstrap
    Acting as bootstrap lead
    Go ahead with the multi-node bootstrap on the rest of
    cluster nodes.  This involves the following commands:

        bootstrap> multi-node
        bootstrap.multi-node> mgmt_ip <SOME_IP>
        bootstrap.multi-node> join 192.168.2.1

        Waiting for joining nodes to join..

At this point, we can go to other nodes and start the bootstrap script,
followng a multi node bootstrap and joining the lead above:

    bootstrap> multi_node
    bootstrap.multi-node> mgmt_ip 192.168.2.2
    This Cluster node's management address will be: 192.168.2.2
    bootstrap.multi-node>
    bootstrap.multi-node>
    bootstrap.multi-node>
    bootstrap.multi-node> join 192.168.2.
    Joining cluster with IP 192.168.2.2, following lead node: 192.168.2.1

At this point the lead node will display:

    \o/ we have a new member claiming mgmt IP 192.168.2.2
    Shall we accept this node into the cluster? [y/n] y
    Member accepted: 192.168.2.2

The joining node will now show:

    Joined cluster.  Awaiting signal from lead to start cluster

Other nodes can follow the same join procedure to join the cluster.  On
this example, we'll assume that another node with IP 192.168.2.3 joined
the cluster.

After sufficient nodes are accepted, the lead bootstrap will display:

    The MidoNet Cluster is READY, members: 192.168.2.1, 192.168.2.3, 192.168.2.3
    Form cluster? [y/n]

If [y] is chosen, the lead node will transmit the final set of nodes to
all the participants and instruct them to start the cluster.

Change-Id: I29229a2db8179b32ec28db9e1eb3a90e03e835fb
Signed-off-by: Galo Navarro <galo@midokura.com>