TraceRequest model for Zoom
This adds a trace request object type for zoom, and a mapper,
TraceRequestChainMapper. The mapper is a trait to be mixed in with any
device which can be traced. When the device data is loaded, if any trace
request ids are set, a trace chain is generated and set as the inbound
filter. The trace chains are shared between the devices and the chain
mapper using a map provided to both by the virtual topology.

Ref MNA-684

Change-Id: I48b527b27675a78f8f3cd9f8cac249b1a9a7b226
Signed-off-by: Ivan Kelly <ivan@midokura.com>