API: Resource filter methods do not return functions
This patch cleans the resource classes from the compatibility API
such that the filter methods are applied directly, rather than
returning Function instances.

The patch is also used in preparation for allowing the filter
methods to return operations that could be chained, simplifying
the write operation for resources that require multiple writes.

Change-Id: Ic35be5a7b03662ec8de82f620b265d2da61a9202
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>