Use a new JVM for every 10 tests
Some tests consume resources which are not freed when the test
finishes. For example, any test that creates a BridgeMapper but doesn't
clean it up will leak the virtualtopology, the injector and anything the
injector had allocated, as the bridge mapper starts a scheduled task and
only cleans it up when the bridge is deleted.

Also, cassandraunit doesn't clean up after itself.

Any memory or direct buffers held by these objects is then held for the
rest of the test run, causing the test jvm to ultimately use many
GB. If direct buffers are not freed, eventually they will
run out.

This change makes a single JVM persist for only 10 test, after which all
resources which have been allocated will be freed.

Change-Id: Ieb7bd493780828b449c033d3cb664ea8d1eeea3d
Signed-off-by: Ivan Kelly <ivan@midokura.com>