API: Improve listing performance for sub-resources
Previously, when listing a sub-resource, such as the ports of a
bridge, the API would fetch all ports from storage and then filter
the resulting list using the `listFilter`, which may be specific to
each resource.

This change introduces a `listIds` filter that is applied before
fetching the objects from storage, and allows a specific list
implementation to return the set of identifiers that have to be
fetched from storage.

Change-Id: I40b22a2989a5d0c5d71910bb45d853b257dc66d0
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>