Maintain a list of traversed devices for flow history
Until now, flow history has been collecting the device tags from the
flow tags. However, these can can contain duplicates and ports which
have not been traversed if the simulation generates flow state (each of
the ports it is sent to is added to the flow tags again). The simplest
solution to this is to store the traversed devices in a separate
list. During simulation, a devices calls addSimulatedDeviceTag. This
will add it to the device list and also to the flow tags.

Ref MN-3730

Change-Id: I436b7aa92cfba9e89d8ab0ad81a4ba8d70c357e7
Signed-off-by: Ivan Kelly <ivan@midokura.com>