Update patch set 1
Patch Set 1:

> I don't like it that this is leaking all over the simulation. This
 > seems like something that should be taken care of in the flow
 > history side of things. 
Yup, not too fond of that myself. The problem is that we need to be able to separate devices which we traversed in simulation, and devices we sent flow state too. This could be done only in flow history, but it would be very hacky, and possibly computationally complex. We'd need to take all tags, which are devices tags, up until all port tags which are in outPorts have been seen. Checking the tags in outPorts on an ArrayList is O(n), so the full check is O(n^2). Now if a bridge has 1000 ports...

 > By removing these devices we're also losing information about
 > which ports we replicated the state to.
Flow history doesn't care about these for now.

Patch-set: 1
Label: Code-Review=0