Update patch set 2
Patch Set 2:

> i don't have a good alternative. i would think though that the flow
 > history could take the list of tags, understand what devices the
 > packet traversed, understand if there were port groups involved,
 > and understand what ports of the port group (the ones that were
 > active) we sent state to. we do preserve order, so these things can
 > be inferred. but maybe i'm misunderstanding the whole feature :p

This understanding is causing flow history to assume a lot of implicit information about how simulation works and requiring a lot of topology queries to build the understanding (which can't be cached well because they rely on the alive status, though zoom should make it better). Also the few of from the flow history node and the view from the agent may differ, throwing a large spanner in the works. And if our debugging tools are inaccurate, they are not very useful.

Patch-set: 2