Update patch set 2
Patch Set 2:

> There's something I'm not getting from your "topology queries"
 > remark. If flow history (FH) is querying ZK - a mutable data store
 > - after it receives a flow, then it cannot be accurate by
 > definition. I'd imagine FH needs to have it's own, log-style view
 > of the virtual topology and can map flows to a given point in the
 > topology's history?

If we record the devices as we traverse them in simulation, we know for sure which devices were traversed in simulation, even if we don't have the devices details available from the metadata store. If we need to simulate the simulation to get this list from a superset of devices, we can't be sure that our result is the same as the true set traversed by the simulation.

Patch-set: 2