Update patch set 2
Patch Set 2:

Hi Jan, we do install the dependencies at run-time but not at this stage of the development process (dependencies are pre-installed at the gates so they are faster). Once the patch is merged, we test a full openstack installation which pulls all the packages from scratch before going to the next step (QA/SA) if that's your concern.

Patch-set: 2