Update patch set 2
Patch Set 2:

> Right, I'm not suggesting any other approach. We should send all
 > devices traversed. Flow history should figure out what devices they
 > are, and optionally prune port groups. Queries should never hit ZK
 > though, or they may be looking at new topology devices.

It's all the ingress port and all egressPorts that get added at that stage, regardless of port groups. 

I did consider using a pruning approach for a while. Take from the tag list until you have seen all outPorts, then next time you see a port disregard it. I didn't like it for two reasons though. Firstly it's a tricky to do in a allocation free way while still keeping complexity low (it's simple with a HashSet). That said we could do this in the post processing phase that isn't in jvm.
Secondly, it implicitly relies on behaviour of the agent to stay the same, making it brittle in the face of future changes. This was what bothered me most.

Patch-set: 2