Agent: Don't remove already removed flows
This patch fixes a bug caused by using a ManagedFlow's refcount to
detect whether it is still alive or not. In particular, the
FlowExpirationIndexer tried to delete a flow if its refcount was
bigger than 1. However, that happens when the flow removal is
in-flight. So, if a flow is expired while it's deletion is in-flight,
the flow refcount is 2 and the FlowExpirationIndexer will issue
another flow deletion request.

Change-Id: I088f34d74ca249583e9bb1067938b19f7d88c82d
Signed-off-by: Duarte Nunes <duarte@midokura.com>