Update patch set 3
Patch Set 3: Code-Review-1

(3 comments)

Hello Alex;

As can be seen from this test:

https://github.com/midonet/midonet/blob/824c2c580c79e78ffe387bff333fcced87e88dd0/nsdb/src/test/scala/org/midonet/cluster/data/storage/StorageTest.scala#L171

As for population of back-reference, when creating inbound & outbound chains for Router, for example, we populate the Router with deterministically generated inbound & outbound chain IDs and then create corresponding chains, and the back-references to the Router are populated with those chains by field binding accordingly, so it doesn't need to be done manually.

Deleting the chain back-reference in Network, Port, etc. can be done by field binding. Field binding cannot be used, at least at the moment, for jump rules, but it seems we should handle the clean up related to jump rule / chain at the translator level.

Patch-set: 3
Label: Code-Review=-1