MDTS: preparation work for switching to sandbox
Moving the start and stop sequence of mmm into scripts. This will be
changed in a later patch that does the actual switch to sandbox
(removing mmm completely and porting mdts to sandbox), so these scripts
will launch sandbox instead of mmm. This way, we'll be effectively
making the switch in one step without touching the gates/nightly jobs.

Change-Id: I768e6e78893b1efb6741d3426df5c98a01b6bbbd
Signed-off-by: Xavi León <xavi.leon@midokura.com>