VXGW: VTEP data methods return futures
The VTEP data client uses an underlying VTEP connection and data to
expose read/write functionality to the VTEP. Currently the data
client await for the VTEP connection to be established before
creating an inner data client that caches the OVSDB tables exposing
the VTEP data. However, the data client public methods are
synchronous, throwing an exception if the connection is not yet
ready, or awaiting for the inner client to make the data available.

This change converts these methods to asynchronous exposing futures
that complete successfully when both the connection is established
and the data is available, simplifying the manner these methods are
used.

Ref: MNA-127

Change-Id: I8d0e756d539401349638dc33b9ab4f14fbd28d26
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>