API: Introduce pluggable resources
This enhances the MidoNet API cluster minion by providing the
capability to load API resources from the current class path at
start time. To this end, it introduces a new `ApiResource`
annotation, and in order for a class to qualify as an API resource
it must be annotated with both the `ApiResource` annotation and
a `Path` annotation indicating the resource name in the URI path.

At the moment, only root Midonet API resource are pluggable.

The `ApiResource` annotation takes an optional `version` argument.
This allows newer resources to override resources with the same
name.

This pluggable resourc model allows to extend the current Midonet
API with resources provided separately from its minion.

Change-Id: I761e80373772b314155d806bb85bf226bc289d75
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>