Cluster API: support https
This patch exposes the MidoNet API via HTTPs.

In mn-conf, cluster.rest_api includes a new property https_port, if this
value is <= 0 the https endpoint will be disabled.  If a valid value is
provided, the API will expect a keystore containing the certificate to
live in /etc/midonet-cluster/ssl/midonet.jks.  This location can be
overriden by setting the midonet.keystore_path system variable.  If the
certificate is not present, HTTPS will be disabled.

To generate a self-signed key, you can use the following procedure. Note
that you'll be prompted for passwords during this process, we'll need to
keep the keystore password for later.

    openssl genrsa -des3 -out midonet.key 2048
    openssl rsa -in midonet.key -out midonet.key
    openssl req -sha256 -new -key midonet.key -out midonet.csr -subj '/CN=localhost'
    openssl x509 -req -days 365 -in midonet.csr -signkey midonet.key -out midonet.crt

Combine private key into cert, because we generated them separately.

    openssl pkcs12 -inkey midonet.key -in midonet.crt -export -out midonet.pkcs12

Load the certificate into the keystore

    keytool -importkeystore -srckeystore midonet.pkcs12 -srcstoretype PKCS12 -destkeystore midonet.jks

Now place the keystore in the default location:

    mv midonet.jks /etc/midonet-cluster/ssl/

The gradle runner works like this:

    ./gradlew cluster:midonet-cluster:run \
        -Pconf=<PATH_TO_CONF_FILE>
        -Pkeystore_path=<PATH_TO_KEYSTORE>
        -Pkeystore_password=<PASSWORD>

Note that the password may be in clear form, or obfuscated.  If you
decide to obfuscate, please refer to [1]

    $ export JETTY_VERSION=9.0.0-SNAPSHOT
    $ java -cp lib/jetty-util-$JETTY_VERSION.jar
    org.eclipse.jetty.util.security.Password

    Usage - java org.eclipse.jetty.util.security.Password [<user>] <password>
    If the password is ?, the user will be prompted for the password

The JETTY_VERSION can be inspected in the root build.gradle.  The jar
should be in your local gradle cache.

Later patches will provide a better way of obfuscating passwords for
end-users.

[1]: https://www.eclipse.org/jetty/documentation/current/configuring-security-secure-passwords.html

Fix: MNA-329
Change-Id: I4fa396d7982aaac368a3447b8118089a7fc7b541
Signed-off-by: Galo Navarro <galo@midokura.com>