Midonet Topology migrator (1.9 -> 2.0)
Migrator for low-level Midonet topology (ignores Neutron) from 1.9 to
2.0. This is a standalone tool that only requires access to Zookeeper;
the Midonet API and Agent don't need to be running.

Migration of BGP- and VTEP-related topology objects is NYI, and
migration of L4LB topology objects is untested. Testing in general is
fairly light so far, so there may be some issues in other areas as
well.

Signed-off-by: Brandon Berg <bberg@midokura.com>
Change-Id: I46b082130673d6c7135942400071fc919da67956