Lock ZK when updating port while binding
In port binding of ZOOM, mm-ctl should lock ZK so that port node does
not get modified while it is doing binding and unbinding.  In the v2
ZK layout, because port binding occurs on the update of the port
object, and since there are many other Neutron translations that modify
the port, concurrent update of the port occurs freqenty.

Fixes: MNA-755
Change-Id: Ibe015ea9e18c453aae14fbe8710c0a0b034d4f67
Signed-off-by: Ryu Ishimoto <ryu@midokura.com>