Generalize BgpdProcess to QuaggaEnvironment
This patch just changes the naming of a few symbols. It does not change
the behavior of anything. The reason why this patch is necessary is that
in future patches, we will treat the bgpd process idea in a more general way.
This new way will include the container running bgpd as well as the zebra
process. In even later patches, we will include also bfdd, ospfd, etc.

As an example, this patch renames vty to bgpVty. When support for bfd is
added, the QuaggaEnvironment will also have a bfdVty, and so on.

Change-Id: I2fdf1746efd8c88f37ee1c005f618f9a2746dbe9