Add FPM server
This commit adds a class that can act as an FPM server for
quagga's FPM feature. This quagga feature will push FIB updates to any
server that is open on a given address and port. The FpmServer will
take these updates, parse them and distill them down to just the info
necessary and sufficient for midonet to handle the route changes.

This patch does not actually affect the behavior of midonet in any
way, it simply adds the functionality that can be used in later patches.

Change-Id: I8fbf6b1c704a411662109ca79ed6854af13fe575