Update patch set 3
Patch Set 3:

Taku,

thank you for comments.
in this case, the consumer of the buffer,
that is, the function specified to flatMap, is executed
in the context of notifications.connect,
as a part of replaying items in ReplaySubject.
at least it's my reading of rxjava.
how do you think?

Patch-set: 3
Reviewer: YAMAMOTO Takashi <1002897@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0