Add FpmServer integration test
This patch introduces an integration test to test the FPM server
functionality. The details of the test are described in detail in
the test file itself, but at a high level the test will create an
FPM server, create a real quagga topology, and then ensure that
expected notifications are received properly.

Change-Id: I085f67b6110afe7d973b88eff8e1fa9cbacef100