VxlanPortService supports new storage stack
This patch modifies VxlanPortMapper such that it supports the new
storage stack. It is renamed to VxlanPortService since it now comes in
the form of a service that can be started and stopped, instead of the
previous actor.

The new implementation follows the new models and VTEP and VTEP
Bindings, and uses the reactive interface exposed by the MidonetBackend.

Ref: MNA-127

Change-Id: I1d75a7435fc1f234e6e3efae29cf4673d535635a
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>
Signed-off-by: Galo Navarro <galo@midokura.com>