Remove legacy API code.
VSphere code moves to cluster along with its tests, although it's still
pending refactor into V2.  The packages for midonet-api are not removed
because they require changing the Sandbox used in MDTS which is a larger
change.

Start to deprecate the flags to run v2 on-demand. From now on, testing
works on v2 only.  Some leftovers still in agent that will need cleanup.

Ref.: MNA-597
Change-Id: I3f736d03e340e3ddc8669ec319e31234b9bf6791
Signed-off-by: Galo Navarro <galo@midokura.com>