Update patch set 4
Patch Set 4:

Sorry, I'm currently working on another task and I need some time to take a look at this deeply.

But let me note that if we create a new buffer for every notification, we'd experience the raise of certain amount of memory consumption, which would not be cheap in my guess because the kernel keeps sending notifications quite frequently and some of them are totally useless or duplicated surprisingly.

So I just want to be careful for this modification and make it sure if it really the root cause of the race.

Patch-set: 4
Label: Code-Review=0