SG rule translator allows open ranges
This fixes an issue in the neutron security group translator,
where open port ranges received for a rule condition via the REST
API were converted into closed ranges with the open bound set to
zero. This caused the ranges to be stored incorrectly in ZOOM,
and any subquent read of the range that checked its correctness
to fail.

The patch also enables the SecurityGroupRuleTranslatorTest and fixes
two tests that were failing because updating a chain when changing
a rule is no longer required.

Fixes: MNA-754

Change-Id: Id4b4e1392e7ebba2d01bf78d69894f2270682139
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>