Update NeutronPort's device owner on RIF create
It turns out that Neutron will sometimes create a port with no device
owner and then create a RouterInterface using the port, instead of the
usual workflow, which is to create a port with device owner set to
ROUTER_INTERFACE and using the for the router interface. When that
happens, RouterInterfaceTranslator needs to update the NeutronPort's
device_owner field to ROUTER_INTERFACE, so that PortTranslator knows
how to handle deletion of the port.

Signed-off-by: Brandon Berg <bberg@midokura.com>
Fixes: MNA-766
Change-Id: I7729ba2d97530ddd34139988333da784bf87d8dc