WIP: VxGW: implementation on the new storage stack
KNOWN FLAKY TESTS PENDING

This class adds a replacement for the VxGW service that works against
the new storage stack.

This also removes the limitation of having a single node doing all vxgw
syncing, and instead distributes the responsibilities among all cluster
nodes.  Each sync process will take care of synchronizing a given VTEP
with all the networks that it has been bound to.

On this patch, since the feature is still disabled, we simply add the
management classes, but NOT implement the load-balancing logic, which
will come in the next patches.

Ref.: MNA-127
Change-Id: I949da289d8013acad8953c0cc2d8126809f3fd50
Signed-off-by: Galo Navarro <galo@midokura.com>