Added Service Chaining models compatible with Neutron SFC API.
Service - a network function accessible via a port. It may be either L2, trans-
parent to traffic, or L3 in which case it only processes packets with its own
MAC in the Ethernet destination, and re-transmits them with its own MAC in the
Ethernet source and its next hop gateway's MAC in the ethernet destination.

Service Chain - a sequence of network functions through which packets should be
directed. The Chain has a name, a port at which it should be inserted, and a
set of traffic classifiers specifying which traffic should be directed through
the chain.

Service Chain Element -

Change-Id: I8333034d81fa136feeebc5175f78f2b1c81f58c5
Signed-off-by: Giuseppe (Pino) de Candia <gdecandia@midokura.com>