Remove VxGW v1 code and odl-ovsdb module
This patch is required so that we stop depending on the old ovdsb
client, which was a fork of the Opendaylight version from early 2014.
We integrate the cluster/vtep module into midonet-cluster, bringing the
upstream plugin from the ODL Lithium release.

In order to allow this, we need to remove all deps on the old plugin,
and hence most of the code needs to be deleted.  The new implementation
based on v2 will be coming in following patches.

Ref.: MNA-127
Change-Id: I5405d80c067dcd26e69dd612c2fea7937ca68db7
Signed-off-by: Galo Navarro <galo@midokura.com>