VxGW: load balance VtepSync among cluster
This patch introduces the Snatcher, a general purpose utility class that
allows watching a given entity from storage, and acquring ownership
based on a StateKey.  Several nodes can compete for ownership and
trigger customizable side effects when the ownership is acquired or
lost.

The Snatcher is used on this patch to incorporate a load balancing and
failure handling strategy into the VxLAN Gateway Service applied to a
VTEP.  Each Cluster node will fight with the rest to control each VTEP.

Ref.: MNA-127
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: Ie6eae25b6e083259c4bda00e0f7478bd04bde666