Agent: Preserve more state when branching
This patch ensures we also preserve the in and out ports at the
current stage in the simulation, since some devices use them.

One example is the BridgePort, which uses the inPort field to drop
packets that a Bridge floods to its input port.

This works right now because we only flood to exterior ports and
because the flood action appears before the ToPortAction to a
vlan-aware bridge port, but for generality we should preserve this
state.

Change-Id: I118827dc44f2b45632bd3093f92e8a1dcc22828f
Signed-off-by: Duarte Nunes <duarte@midokura.com>