Support client and pool members in the same subnet
In the current L4LB implementation, midonet does not load balance
properly if the sender of the traffic and the backend servers are in
the same subnet.  This patch adds SNAT rules in the router chain to
support this case.

This patch contains a bug in which two router interfaces on the same
router are not properly handled. The following patch will fix the bug.

Change-Id: I4f3d808a5c72180b4db3a4725f25c7c2807147be
Partially Fixes: MN-3812