Update patch set 2
Patch Set 2: Code-Review+2

Thank, Joe. This looks good as is and feel free to merge it after it passes the Jenkins. If you're willing to add tests, then can you add a similar test to this?

https://github.com/midonet/midonet/blob/master/cluster/midonet-cluster/src/test/scala/org/midonet/cluster/services/c3po/translators/PortTranslatorTest.scala#L1197

You can almost copy and past most of the test code.

Patch-set: 2
Label: Code-Review=+2