Deterministicly set router port id for haproxy
When setting up an haproxy namespace, we create a router port to connect
the namespace to midonet. However, we first check to see if that port
exists before creating it, and just use it if its already present. This
is incorrect behavior because we currently have no way to distinquish
between different namespace's router port, and therefore the most
recent namespace creation will simply take over the port from
whoever was currently using it.

This patch fixes this problem by using a deterministicly generated id
based on the pool id. This way we can always get the router port id
for any given pool.

Change-Id: Ief240ef5314949e7c65f03352b9486a59f1fb56b