VxGW: bindings manage vxlan ports
This patch completes the vtep binding rest API so that vxlan ports are
created in the networks that first get a binding to a VTEP, and removed
when they lose it.

Tests are not added because they are already written on the midonet-api
module, but they need adaptation to use the new mock.

Ref.: MNA-127
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: Ic37ad4ebdabdfad8daeac8afd1e404f9fb7cabbe