Multiple inbound/outbound filters for devices
Allow devices to receive a list of inbound/outbound filter ids rather
than a single ID. The device checks packets against these filters in
the same order as they are provided in.

This change lays the ground work for service insertion. It also
simplifies trace chains, by removing the need to have a jump rule at
the end of the trace chain.

Change-Id: I7ad96a4c3866143c8021dde144e88ceb2d24af88
Signed-off-by: Ivan Kelly <ivan@midokura.com>