Add L2Transform rule type, and redirect action
Part of l2insertion. Add a new rule type, L2TRANSFORM, which can push
and pop vlan. Also adds a new rule action, REDIRECT, which hijacks a
packet and pushes it to another datapath port.

A redirect can either ingress or egress a port. In the case it egresses
a port, all outfilters are skipped.

Change-Id: I28685e12dd9b5379b4f5531539c446da56eb70fd
Signed-off-by: Ivan Kelly <ivan@midokura.com>