VXGW: Improvements to the VTEP client
The patch moves the OVSDB transaction wrapers to an OvsdbOperations
class. It enhances these operations to support completion with the
NETTY channel close future, and simplifies the code with Scala. Now
all operations return Scala futures allowing them to be composed in
a sequential manner.

The changes also accommodate the behavior of the actual OVSDB
database, by changing the way responses are handled and translated
into results.

Finally, it ensures that the cached table monitor receives updates
by setting the monitor select to true for any initial, insert and
delete updates.

Ref: MNA-127
Ref: MNA-800

Change-Id: I3990163452b87c6e40567f7a83b39a1745d9f6fd
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
Signed-off-by: Galo Navarro <galo@midokura.com>