MDTS: Redirect logging information to a file
Quick fix to redirect logging information to a file so can be inspected
later on the CI. It's not ideal as all output from all tests are shown
in this file. Ideally, the log output should be split by failed test
case. Logcapture plugin should do this but not working as expected.

Change-Id: Ifef1269fa0ebcab9cab99cdb06b43340ac71b97e