API: VendorMediaType cleanups, fix Condition
In favour of the newer MidonetMediaTypes.  This also keeps all the last
versions of each media type, and updates tests accordingly.  Since and
Until annotations are also removed, as they are not applicable anymore.

This patch uncovers a bug in the TraceRequests API where Conditions are
not deserialized properly from ZK.  This was because the Proto was
setting a default value in mask fields, which was deserialized into a
non-null value on the DTO.

Given that fixing the bug is a requisite to merge the patch, we're doing
it here, enabling TestTraceRequest unit tests.

Change-Id: I85dc34353e5ca73b801f46f30a3f53058b35c3cc
Signed-off-by: Galo Navarro <galo@midokura.com>