Update patch set 2
Patch Set 2:

One thing I noticed with the nose log was that it pulled in everything from the midolman logs, the zookeeper logs, cass logs etc. Since these are already being put in the tarball, they can be left out of the noselog. It would also make it quicker to open the file.

Patch-set: 2
Label: Code-Review=0