OVSDB: connection mock + enable VTEP tests
This patch refactors some of the code around the OvsdbVtepConnection so
that we can inject mocked instances in tests, taking the chance to
elliminate some redundant/unneeded abstractions.

We integrate the mock into the API VTEP tests by enabling some of the
disabled tests that use the mock.

The mock VTEP connection will accept any management IP as valid, and
fake an OVSDB that contains a pre-set list of ports available for use in
tests.

We also enable the majority of API tests for VTEP.

Ref.: MNA-127
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: If2935ef2fb97b54ec6e140c11aa7ba9cf3d64d7c